+++ doc/phase1/let/i16-max/expected.c
@@ -3,6 +3,6 @@
 int main(void) {
     TSC_INIT();
     int16_t x = 32767;
-    printf("%d\n", x);
+    printf("%d\n", (int)x);
     return 0;
 }
