+++ doc/phase1/let/i16-min/expected.c
@@ -3,6 +3,6 @@
 int main(void) {
     TSC_INIT();
     int16_t x = -32768;
-    printf("%d\n", x);
+    printf("%d\n", (int)x);
     return 0;
 }
