+++ doc/phase1/let/i8-min/expected.c
@@ -3,6 +3,6 @@
 int main(void) {
     TSC_INIT();
     int8_t x = -128;
-    printf("%d\n", x);
+    printf("%d\n", (int)x);
     return 0;
 }
