+++ doc/phase1/let/u16-zero/expected.c
@@ -2,7 +2,7 @@
 
 int main(void) {
     TSC_INIT();
-    uint16_t x = 0;
-    printf("%u\n", x);
+    uint16_t x = 0U;
+    printf("%u\n", (unsigned)x);
     return 0;
 }
