+++ doc/phase2/utility-types/number-to-fixed/expected.c
@@ -3,7 +3,7 @@
 int main(void) {
     TSC_INIT();
     char _buf_0[64];
-    snprintf(_buf_0, sizeof(_buf_0), "%.2f", 3.14159);
+    snprintf(_buf_0, sizeof(_buf_0), "%.2f", (double)3.14159);
     const String x = STR_LIT_RUNTIME(_buf_0);
     printf("%s\n", x.data);
     return 0;
